ft_transcendence/apps/djangoproject/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
distlib==0.3.8
Django==3.2.5
django-cors-headers==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2-binary==2.9.3
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
